[WIP] Update Dikit

+++ MiniTwo/Core/Data/Entities/Plan.swift
@@ -10,21 +10,21 @@ import Foundation
 class Plan: Identifiable, Equatable {
     let id: String = UUID().uuidString
     var name: String
-    var propertyPrice: Decimal
-    var downPaymentAmount: Decimal?
+    var propertyPrice: Double
+    var downPaymentAmount: Double?
     var tenor: Int
     
     var fixInterests: [Float]
     var floatingInterests: [Float]
     
-    var pokok: Decimal = 0
+    var pokok: Double = 0
     
     var bankProgram: BankProgram?
     
     init(
         name: String,
-        propertyPrice: Decimal,
-        downPaymentAmount: Decimal? = 0,
+        propertyPrice: Double,
+        downPaymentAmount: Double? = 0,
         tenor: Int,
         
         fixInterests: [Float] = [],
@@ -46,9 +46,9 @@ class Plan: Identifiable, Equatable {
     
     init(planEntity: PlanEntity) {
         self.name = planEntity.name ?? ""
-        self.downPaymentAmount = planEntity.dp as Decimal?
+        self.downPaymentAmount = planEntity.dp?.doubleValue
         self.tenor = Int(planEntity.tenor)
-        self.propertyPrice = planEntity.propertyPrice! as Decimal
+        self.propertyPrice = planEntity.propertyPrice!.doubleValue
         
         self.fixInterests = []
         self.floatingInterests = []
